Remove independent wire test
